LayoutTests/platform/ios-wk2/fast/scrolling/scrollIntoView-on-option-expected.txt


FAIL scroller.scrollTop should be within 10 of 790. Was 0.
FAIL listbox.scrollTop should be within 2 of 52. Was 0.
PASS successfullyParsed is true

TEST COMPLETE
